Computer-Vision-Practicals/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
pytorch-lightning>=2.2.0
transformers>=4.40.0
timm>=1.0.0
accelerate>=0.27.0


opencv-python>=4.8.0
albumentations>=1.4.0
segmentation-models-pytorch>=0.3.3
Pillow>=10.0.0


ultralytics>=8.0.0
python-doctr[torch]>=0.9.0
mineru>=0.1.0


numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.65.0
PyYAML>=6.0
rich>=13.0.0


streamlit>=1.32.0


wandb>=0.16.0


ipython>=8.0.0
